Smart(er) exit handeling, verbose added
Added the verbose command line option and moved
the printElements and printReactions calls within
a verbose if-statement.  Exiting the program early
will no longer cause ncurses to write junk all
over the terminal screen.  maxIters is adjusted in
the config.out file after an early exit, and if
atomCount is too high to fit all the atoms into
the world, it will reduce the value to worldX *
worldY and reflect that in the config.out file.
Progress updates are now printed at short, regular
real time intervals. The short name for --sleep
was changed from -l to -z ("zzzz...").