Major performance improvements, progress option
The -p, --no-progress option has been added for
disabling progress reporting.  The usleep function
was wasting time even when o->sleep=0, so the call
is now wrapped in an if-statement.  Also,
o->useGUI now refers to all ncurses functions, and
a cout version of the verbose content and progress
output is used when the gui/ncurses is disabled.
The combination of ncurses functions move(y,x) and
getch() were nearly doubling runtime, even with
printWorld disabled (and only when both move and
getch were present, although getch contributes
to the inefficiency somewhat on its own
regardless).