Reactions no longer must preserve atom count
Elements now have character symbols in addition to
their string names.  To allow for Reactions such
as A -> B + C, the Element type "Solvent" now
exists with symbol '*', and such Reactions are
modelled as A + * -> B + C.  However, to keep the
simulation running as quickly as possible, empty
world spaces are not filled with Solvent Atoms
and continue to be NULL pointers.  Nevertheless,
empty spaces are always considered for Reactions
in executeRxns.