Script updates and config.out note for shuffling
Added a note as to whether shuffling was turned on
or not in the config.out file. Updated
rate_batch.sh script to make it very easy to turn
shuffling on or off. Updated R analysis scripts
accordingly, and improved some plot axes labels
and titles. Also added support for zeroth-order
reactions to rate_analysis.R. rate_analysis_mm.R
now ignores the first 50% of iterations for
determining initial velocity to allow for a steady
state to develop.