Generalized printRxns and printEles to use streams
printElements and printReactions were renamed
printEles and printRxns.  They now take an ostream
pointer as an argument.  Printing is directed to
this stream.  If a NULL pointer is passed, the
ncurses printing functions are called instead.
A printInits function was added to match.
writeConfig now takes advantage of all of these.
printRxns also writes properly formatted reaction
equations which work again with --load.  Need to
update README.