Beginnings of Qt support
makefile and .pro file are now integrated to allow
for the simulator to be built with or without Qt
libraries and with or without ncurses libraries.
"gui" option now has three internal settings,
although the Qt gui option setting cannot be
activated yet.  Qt is not actually used for any
visualization yet.