Command line options are GUI-aware
Command line options added for switching between
GUIs. --gui-ncurses is available only when Qt and
ncurses are both included in the compilation.
--gui-off is included if either Qt or ncurses are
included in the compilation.  The GUI is set to Qt
by default if available, to ncurses if ncurses is
available but Qt is not, or to off if neither are
available. The GUIs supported in a particular
compilation are listed under --version. Also
removed short-opt forms for --rxns-on and
--shuffle-off, renamed the long-opt names for -x
and -y to --width and --height, and updated the
--help.