Added metabolism-debug; make profile uses gprof
Created a new build target called 'make
metabolism-debug' which enables the '-g' and '-pg'
flags for debug info and profiling info. The
target is otherwise identical to 'make
metabolism-minimal' and does not include Qt or
ncurses. The 'make profile' target takes advantage
of this new build and now uses gprof instead of
oprofile. Profiling output is dumped to
profile.out.