"extinct" keyword added for config files
The "extinct" keyword allows users to specify
conditions for early termination of the
simulation. All load files were updated, and
consequently the 'check' files needed to be
updated. README was updated. A census is now taken
for the final state of the lattice before
terminating, even if a census was taken recently.