Qt prompt for saving output files added
When the Qt gui is used, any --files options are
ignored. Instead, dialog boxes will guide the user
through the saving process when the app is closed.
Files are saved in a temporary location until the
user confirms that they should be saved
permanently. Temporary files are always deleted
when the app exits.