README example for Mac finally working properly
It is no longer recommended that Mac users use
'open' for starting the program when attempting
to --load a config file. 'open' "forgets" the pwd
and assumes all paths are relative to the top-
level directory "/". Also, a new parameter was
recently added for rate_plot.R, and so the command
for creating the plots in the example did not work
because it was out of date.  This is fixed. The
example runs as expected on dynamicspjt.