Element starting concentrations implemented
Instead of using a maxAtoms count and an initTypes
list to randomly determine an Atom's type when
created, the randomized positions array is divided
into 8 equally sized sets of positions, and each
Element is assigned one set for placing Atoms. The
value 8 can be changed at compile time in sim.h.
This places a cap on the number of Elements that
can exist in the periodicTable at any one time.
The starting concentration value for each Element
is used to determine what proportion of the set of
lattice squares should be filled. This change was
implemented in preparation for improvements to the
gui control panel. The --atoms option has been
removed, as has the init keyword for config files.